+++ client/js/components/procedure/StatementSegmentsList/StatementSegmentsList.vue
@@ -598,25 +598,6 @@ export default {
         })
     },
 
-    fetchCustomFields () {
-      const payload = {
-        id: this.procedure.id,
-        fields: {
-          AdminProcedure: [
-            'segmentCustomFields',
-          ].join(),
-          CustomField: [
-            'name',
-            'description',
-            'options',
-          ].join(),
-        },
-        include: ['segmentCustomFields'].join(),
-      }
-
-      this.getAdminProcedureWithFields(payload)
-    },
-
     getActionFromQueryParams (queryParams) {
       const action = queryParams.get('action')
 
@@ -822,9 +803,6 @@ export default {
 
     this.setReturnLink()
 
-    if (hasPermission('field_segments_custom_fields')) {
-      this.fetchCustomFields()
-    }
     this.listAssignableUser({
       include: 'orga',
       fields: {