+++ client/js/components/statement/originalStatementsTable/OriginalStatementsTableItem.vue
@@ -122,6 +122,8 @@
           class="px-4 mb-4"
           mode="readonly"
           resource-type="OriginalStatement"
+          source-entity="PROCEDURE"
+          target-entity="STATEMENT"
           expandable
         >
           <template v-slot:readonly-display="{ field }">