+++ CHANGELOG.md
@@ -6,6 +6,9 @@
 
 ## UNRELEASED
 
+### Added
+- Support multiple custom field types and target contexts per project
+
 ## v4.37.0 (2026-04-27)
 
 ## v4.36.0 (2026-04-24)